Change db numbers to float

+++ db/migrate/20160216194318_create_category_items.rb
@@ -3,10 +3,10 @@ class CreateCategoryItems < ActiveRecord::Migration
     create_table :category_items do |t|
     	t.references :category_scanner, null: false
     	t.string :category, null: false
-    	t.integer :quantity, null: false
-    	t.integer :retail_price, null: false
-    	t.integer :cost
-    	t.decimal :margin
+    	t.float :quantity, null: false
+    	t.float :retail_price, null: false
+    	t.float :cost
+    	t.float :margin
 
     	t.timestamps null: false
     end
